+++ packages/file-util/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@openapi-platform/download-util",
+  "name": "@openapi-platform/file-util",
   "version": "1.0.0-alpha.0",
   "main": "./lib/index.js",
   "types": "./src/index.ts",
@@ -9,6 +9,7 @@
     "adm-zip": "^0.4.11",
     "del": "^3.0.0",
     "fs-extra": "^7.0.0",
-    "globby": "^8.0.1"
+    "globby": "^8.0.1",
+    "node-fetch": "^2.2.0"
   }
 }
