+++ packages/file-util/src/index.ts
@@ -1,9 +1,9 @@
-import fs from 'mz/fs';
 import os from 'os';
 import { join } from 'path';
 
+import AdmZip from 'adm-zip';
 import del from 'del';
-
+import fs from 'mz/fs';
 import fetch from 'node-fetch';
 
 export async function makeTempDir(prefix: string) {
@@ -36,4 +36,18 @@ export async function downloadToPath(localPath: string, remotePath: string) {
     stream.on('error', reject);
   });
   await writeToFilePromise;
+}
+
+/**
+ * SDKs are downloaded as a zip file so they need to be extracted.
+ * Use this method to do so.
+ * @param archiveFilePath The archive file path
+ * @param extractToDir Where you want the archive files to be extracted to
+ */
+export async function extractSdkArchiveFileToDir(
+  extractToDir: string,
+  archiveFilePath: string,
+) {
+  const zip = new AdmZip(archiveFilePath);
+  await zip.extractAllTo(extractToDir, true);
 }
