+++ packages/git-util/package.json
@@ -11,7 +11,7 @@
     "mz": "^2.7.0",
     "node-fetch": "^2.1.2",
     "fs-extra": "^7.0.0",
-    "adm-zip": "^0.4.11",
-    "globby": "^8.0.1"
+    "globby": "^8.0.1",
+    "@openapi-platform/file-util": "1.0.0-alpha.0"
   }
 }
