+++ packages/git-util/src/index.ts
@@ -1,13 +1,18 @@
 import oldFs from 'fs';
-import { relative } from 'path';
+import { relative, join } from 'path';
 
 import { clone, push, add, commit } from 'isomorphic-git';
 
+import { move } from 'fs-extra';
+import globby from 'globby';
+import { listFiles } from 'isomorphic-git';
+
 import {
-  getAllStageableFilepathsInRepo,
-  migrateSdkIntoLocalRepo,
-} from '@openapi-platform/download-util';
-import { deletePaths, makeTempDir } from '@openapi-platform/download-util/lib/file';
+  deletePaths,
+  makeTempDir,
+  downloadToPath,
+  extractSdkArchiveFileToDir,
+} from '@openapi-platform/file-util';
 import { GitInfo } from '@openapi-platform/model';
 
 export async function updateRepoWithNewSdk(
@@ -74,5 +79,82 @@ export async function updateRepoWithNewSdk(
     throw err;
   } finally {
     await deletePaths([repoDir]);
+  }
+}
+
+// Note: dir = directory
+
+export async function getAllStageableFilepathsInRepo(repoDir: string) {
+  return await globby([join(repoDir, '**'), join(repoDir, '**', '.*')], {
+    gitignore: true,
+  });
+}
+
+export async function getAllFilepathsInDir(dir: string) {
+  return await globby([join(dir, '**'), join(dir, '**', '.*')]);
+}
+
+export async function moveFilesIntoLocalRepo(repoDir, sdkDir) {
+  // Unfortunately you have to move each file individualy (to knowledge)
+  const paths = await getAllFilepathsInDir(sdkDir);
+  for (const path of paths) {
+    const fromPath = path;
+    const relativePath = relative(sdkDir, path);
+    const toPath = join(repoDir, relativePath);
+    await move(fromPath, toPath, { overwrite: true });
+  }
+}
+
+async function deleteAllFilesInLocalRepo(dir) {
+  const filePaths = await listFiles({ fs: oldFs, dir });
+  const fullFilePaths = filePaths.map(path => join(dir, path));
+  await deletePaths(fullFilePaths);
+}
+
+/**
+ * Tries to put the files of a remotely stored sdk ZIP file into a locally stored
+ * repository.
+ */
+export async function migrateSdkIntoLocalRepo(
+  repoDir: string,
+  remoteSdkUrl: string,
+  fileCleaningGlobs: string[],
+  options,
+) {
+  const { logger } = options;
+  logger.verbose(`Deleting all files ${repoDir}`);
+  await deleteAllFilesInLocalRepo(repoDir);
+  /*
+   * We make folders for each step of the process,
+   * one for downloading, one for unzipping.
+   */
+  const downloadDir = await makeTempDir('download');
+  try {
+    const sdkArchiveFilePath = join(downloadDir, 'sdk.zip');
+    await downloadToPath(sdkArchiveFilePath, remoteSdkUrl);
+    const sdkDir = await makeTempDir('sdk');
+    try {
+      logger.verbose(`Extracting ${sdkArchiveFilePath} to ${sdkDir}`);
+      await extractSdkArchiveFileToDir(sdkDir, sdkArchiveFilePath);
+
+      for (let glob of fileCleaningGlobs) {
+        glob = join(sdkDir, glob);
+      }
+      logger.verbose(
+        `Deleting files from ${sdkDir} which match globs: ${fileCleaningGlobs}`,
+      );
+      await deletePaths(fileCleaningGlobs);
+
+      logger.verbose(`Moving files from ${sdkDir} to ${repoDir}`);
+      await moveFilesIntoLocalRepo(repoDir, sdkDir);
+    } catch (err) {
+      throw err;
+    } finally {
+      await deletePaths([sdkDir]);
+    }
+  } catch (err) {
+    throw err;
+  } finally {
+    await deletePaths([downloadDir]);
   }
 }
