+++ packages/git-util/test/system/git.test.ts
@@ -18,8 +18,8 @@ jest.mock('isomorphic-git', () => {
   return mockedGitModule;
 });
 
-jest.mock('../../src/file/index', () => {
-  const actualModule = require.requireActual('../../src/file/index');
+jest.mock('@openapi-platform/file-util', () => {
+  const actualModule = require.requireActual('@openapi-platform/file-util');
   const { mockFunctions } = require('jest-mock-functions');
   const mockedModule = mockFunctions(actualModule, {
     onMockedFunction: (fn, ogFn) => fn.mockImplementation((...other) => ogFn(...other)),
@@ -40,7 +40,7 @@ const logger = openapiLogger();
  * This is actually a test to make sure the other tests are going to work.
  */
 it('temp dir actually works', async () => {
-  const { makeTempDir } = require('../../src/file/index');
+  const { makeTempDir } = require('@openapi-platform/file-util');
   await expect(makeTempDir('test')).resolves.not.toBeUndefined();
 });
